chore: move filename parsing to ContentDispositionParser as utility methods
